// src/app/model/kpi.interface.ts
// export class Ikpi {
//     // name: string;
//     icon: string;
//     value: number;
//     subtitle: string;
//     // useracc: number;
//     // userapp: number;
//     // mpracc: number;
//     // mprapp: number
//     // total1: number;
//     // total2: number;

//     constructor( icon: string,  value: number, subtitle: string){
//         this.icon=icon;
//         this.value=value;
//         this.subtitle=subtitle;
//     }
// };
export interface Ikpi {
    
    icon: string;
    value: number;
    subtitle: string;
    useracc: number;
    userapp: number;
    mpracc: number;
    total1: number;
    total2:number;
    mprapp:number;
    


}